Add a header to share CPU-specific PECI data
This change moves some of the generic CPU-specific PECI data to
a separate header so that it can be shared with other modules
that need similar capability.

Tested:
Combined with the host-error-monitor changes confirmed that when
an error is injected, the host-error-monitor can successfully
determine the CPU model and make the correct PECI calls to log
the error.

Change-Id: I92168ad3372b5fbfc2bde13f13a4bc9fec870fef
Signed-off-by: Jason M. Bills <jason.m.bills@linux.intel.com>