Make the connection object global
Instead of passing around the connection object, make it global.
This will allow for future capability to log error counts on
DBus without passing in the connection object.

Tested:
Combined with the log count change confirmed that when a new
Crashdump is generated, it can successfully increment the counter
on DBus.

Change-Id: I0ca185d1fb92460babf712da03c56c303833ffc6
Signed-off-by: Jason M. Bills <jason.m.bills@linux.intel.com>