Move PECI CPU info into libpeci
This removes peci_cpus.hpp and moves the relevant information
directly into libpeci.

It also adds a new function in libpeci to get the CPUID, so it
changes crashdump to use that instead.

Tested:
Injected an ERR2 timeout and confirmed that crashdump completed
successfully.

Manually generated a crashdump before and after this change and
confirmed that they match.

Change-Id: I9431a446cc58ba8fc4b8ba1b7dfa13a6d07e0de9
Signed-off-by: Jason M. Bills <jason.m.bills@intel.com>