Fix log interface vector reserve error
Constructing with a count is not the same as calling reserve().
The constructor was adding three blank elements to the vector and
increasing the size.  The desired behavior is reserve() which
allocates memory for three elements but doesn't insert them into
the vector.

Tested:
Checked the initial size of the vector and confirmed that it is 0
instead of 3.

Change-Id: Ifdcaeb2f09fbe3f1f7cde7e23a72bae9b1996f3b
Signed-off-by: Jason M. Bills <jason.m.bills@intel.com>